Tidied up the flow based on feedback
