+++ app/models/wagtail.py
@@ -1,15 +1,12 @@
 from typing import Optional
 
 import time
-import urllib.parse
+import uuid
 from datetime import datetime
-from importlib.metadata import metadata
-from urllib.parse import urlencode
 
 import djstripe.models
 import orjson
 import shopify
-import stripe
 from django.conf import settings
 from django.contrib.contenttypes.models import ContentType
 from django.contrib.postgres.fields import ArrayField
@@ -17,70 +14,42 @@ from django.core.cache import cache
 from django.core.validators import MinValueValidator
 from django.db import models
 from django.db.models import Q
-from django.http.response import Http404
-from django.shortcuts import get_object_or_404, redirect
 from django.templatetags.static import static
-from django.urls import reverse, reverse_lazy
+from django.urls import reverse
 from django.utils.decorators import method_decorator
 from django.utils.html import strip_tags
-from django_countries import countries
 from djmoney.models.fields import Money, MoneyField
 from modelcluster.fields import ParentalKey, ParentalManyToManyField
 from modelcluster.models import ClusterableModel
-from wagtail import blocks
 from wagtail.admin.panels import (
     FieldPanel,
     FieldRowPanel,
     HelpPanel,
     InlinePanel,
     MultiFieldPanel,
-    PageChooserPanel,
     StreamFieldPanel,
 )
 from wagtail.contrib.routable_page.models import RoutablePageMixin, route
-from wagtail.fields import RichTextField, StreamField
-from wagtail.images.blocks import ImageChooserBlock
+from wagtail.fields import RichTextField
 from wagtail.images.edit_handlers import ImageChooserPanel
 from wagtail.images.models import AbstractImage, AbstractRendition
 from wagtail.models import Orderable, Page, Site
 from wagtail.rich_text import get_text_for_indexing
 from wagtail.search import index
-from wagtail.snippets.blocks import SnippetChooserBlock
 from wagtail.snippets.models import register_snippet
 from wagtailautocomplete.edit_handlers import AutocompletePanel
 from wagtailcache.cache import WagtailCacheMixin, cache_page
 from wagtailseo import utils
 from wagtailseo.models import SeoMixin, SeoType, TwitterCard
 
-from app.models.blocks import ArticleContentStream
+from app.models.blocks import *
 from app.models.circle import CircleEvent
-from app.models.django import User
-from app.utils import include_keys
 from app.utils.abstract_model_querying import abstract_page_query_filter
 from app.utils.cache import django_cached
 from app.utils.shopify import metafields_to_dict
 from app.utils.stripe import create_shipping_zone_metadata, get_shipping_product
 
 from .stripe import LBCProduct, ShippingZone
-
-block_features = [
-    "h1",
-    "h2",
-    "h3",
-    "h4",
-    "h5",
-    "bold",
-    "italic",
-    "link",
-    "ol",
-    "ul",
-    "hr",
-    "link",
-    "document-link",
-    "image",
-    "embed",
-    "blockquote",
-]
 
 
 class SeoMetadataMixin(SeoMixin, Page):
@@ -215,6 +184,7 @@ class MembershipPlanPrice(Orderable, ClusterableModel):
         LBCProduct,
         blank=True,
         help_text="(For V1-only signup flow.) The stripe product that the user will be subscribed to. If multiple products are set here, then the user will be asked to pick which one they want, e.g. Classic or Contemporary books.",
+        verbose_name="[V1] stripe products",
     )
 
     panels = [
@@ -272,6 +242,15 @@ class MembershipPlanPrice(Orderable, ClusterableModel):
             return Money(0, zone.rate_currency)
         return zone.rate * self.deliveries_per_billing_period
 
+    def equivalent_monthly_shipping_fee(self, zone) -> Money:
+        if self.free_shipping_zones.filter(code=zone.code).exists():
+            return Money(0, zone.rate_currency)
+        return (
+            zone.rate
+            * self.deliveries_per_billing_period
+            / self.months_per_billing_cycle
+        )
+
     def price_including_shipping(self, zone):
         return self.price + self.shipping_fee(zone)
 
@@ -327,6 +306,20 @@ class MembershipPlanPrice(Orderable, ClusterableModel):
         return self.price / self.months_per_billing_cycle
 
     @property
+    def monthly_price_string(self) -> str:
+        money = str(self.equivalent_monthly_price)
+        s = f"{money}/month"
+        return s
+
+    def equivalent_monthly_price_string_including_shipping(self, zone) -> str:
+        money = str(
+            self.equivalent_monthly_price
+            + self.shipping_fee(zone) / self.months_per_billing_cycle
+        )
+        s = f"{money}/month"
+        return s
+
+    @property
     def equivalent_monthly_price_string(self) -> str:
         money = str(self.equivalent_monthly_price)
         s = f"{money}/month"
@@ -335,7 +328,11 @@ class MembershipPlanPrice(Orderable, ClusterableModel):
         return s
 
     @property
-    def price_string_uk_pp(self) -> str:
+    def shipping_price_string_uk(self) -> str:
+        return self.shipping_fee(ShippingZone.get_for_country(iso_a2="GB"))
+
+    @property
+    def price_string_including_uk_shipping(self) -> str:
         return self.price_string_including_shipping(
             ShippingZone.get_for_country(iso_a2="GB")
         )
@@ -443,6 +440,16 @@ class MembershipPlanPrice(Orderable, ClusterableModel):
             ).first()
         return current_plan_price
 
+    @property
+    def discount_percent(self):
+        if self.interval != "year":
+            return None
+        monthly_price = self.plans.first().monthly_price.price
+        discount_percent = (
+            monthly_price - self.equivalent_monthly_price
+        ) / monthly_price
+        return discount_percent
+
 
 @register_snippet
 class Upsell(Orderable, ClusterableModel):
@@ -488,13 +495,6 @@ class Upsell(Orderable, ClusterableModel):
         )
 
 
-book_types = [
-    ("classic", "classic"),
-    ("contemporary", "contemporary"),
-    ("all-books", "all-books"),
-]
-
-
 @register_snippet
 class SyllabusPage(Page, Orderable, ClusterableModel):
     # title
@@ -513,135 +513,22 @@ class SyllabusPage(Page, Orderable, ClusterableModel):
         help_text="The stripe product that the user will be subscribed to.",
     )
 
-    # plan = ParentalManyToManyField("app.MembershipPlanPage", related_name="syllabi", help_text="The delivery plans that this syllabus is available on.")
+    @classmethod
+    def create_default_layout(cls):
+        return {"id": uuid.uuid4(), "type": "syllabus_title", "value": {}}
+
+    layout = create_streamfield(
+        [
+            ("syllabus_title", SyllabusTitleBlock()),
+        ],
+        default=create_default_layout,
+    )
 
     content_panels = Page.content_panels + [
         FieldPanel("description"),
         FieldPanel("book_types"),
         AutocompletePanel("stripe_product", target_model=LBCProduct),
     ]
-
-
-class PlanTitleBlock(blocks.StructBlock):
-    class Meta:
-        template = "app/blocks/plan_title_block.html"
-
-    # def get_context(self, value, parent_context=None):
-    #     context = super().get_context(value, parent_context)
-    #     context['page'] = parent_context['page']
-    #     return context
-
-
-class PlanPricingBlock(blocks.StructBlock):
-    class Meta:
-        template = "app/blocks/plan_pricing_block.html"
-
-    # def get_context(self, value, parent_context=None):
-    #     context = super().get_context(value, parent_context)
-    #     context['page'] = parent_context['page']
-    #     return context
-
-
-class BookTypeChoice(blocks.ChoiceBlock):
-    choices = book_types
-
-    class Meta:
-        default = "all-books"
-
-
-class BackgroundColourChoiceBlock(blocks.ChoiceBlock):
-    choices = [
-        ("bg-black text-white", "black"),
-        ("bg-white", "white"),
-        # ('primary', 'primary'),
-        ("tw-bg-yellow", "yellow"),
-        # ('black', 'black'),
-        ("tw-bg-teal", "teal"),
-        ("tw-bg-darkgreen", "darkgreen"),
-        ("tw-bg-lilacgrey", "lilacgrey"),
-        ("tw-bg-coral", "coral"),
-        ("tw-bg-purple", "purple"),
-        ("tw-bg-magenta", "magenta"),
-        ("tw-bg-pink", "pink"),
-        ("tw-bg-lightgreen", "lightgreen"),
-    ]
-
-    class Meta:
-        icon = "fa-paint"
-
-
-class AlignmentChoiceBlock(blocks.ChoiceBlock):
-    choices = [
-        ("left", "left"),
-        ("center", "center"),
-        ("right", "right"),
-    ]
-
-    class Meta:
-        icon = "fa-arrows-h"
-        default = "center"
-
-
-class BootstrapButtonSizeChoiceBlock(blocks.ChoiceBlock):
-    choices = [
-        ("sm", "small"),
-        ("md", "medium"),
-        ("lg", "large"),
-    ]
-
-    class Meta:
-        icon = "fa-arrows-alt"
-        default = "md"
-
-
-class BootstrapButtonStyleChoiceBlock(blocks.ChoiceBlock):
-    choices = [
-        ("btn-outline-dark", "outlined"),
-        ("btn-dark text-yellow", "filled"),
-    ]
-
-    class Meta:
-        default = "btn-outline-dark"
-
-
-class PlanBlock(blocks.StructBlock):
-    plan = blocks.PageChooserBlock(
-        page_type="app.membershipplanpage",
-        target_model="app.membershipplanpage",
-        can_choose_root=False,
-    )
-    background_color = BackgroundColourChoiceBlock(required=False)
-    promotion_label = blocks.CharBlock(
-        required=False, help_text="Label that highlights this product"
-    )
-
-    def get_context(self, value, parent_context=None):
-        context = super().get_context(value, parent_context)
-        context["request_price"] = value["plan"].get_price_for_request(
-            parent_context.get("request")
-        )
-        return context
-
-    class Meta:
-        template = "app/blocks/membership_option_card.html"
-        icon = "fa-money"
-
-
-class MembershipOptionsBlock(blocks.StructBlock):
-    heading = blocks.CharBlock(
-        required=False,
-        form_classname="full title",
-        default="Choose your plan",
-    )
-    description = blocks.RichTextBlock(
-        required=False,
-        default="<p>Your subscription will begin with the most recently published book in your chosen collection.</p>",
-    )
-    plans = blocks.ListBlock(PlanBlock)
-
-    class Meta:
-        template = "app/blocks/membership_options_grid.html"
-        icon = "fa-users"
 
 
 class CustomImage(AbstractImage):
@@ -904,310 +791,6 @@ class BaseShopifyProductPage(ArticleSeoMixin, Page):
         )
 
 
-class ButtonBlock(blocks.StructBlock):
-    text = blocks.CharBlock(max_length=100, required=False)
-    page = blocks.PageChooserBlock(
-        required=False, help_text="Pick a page or specify a URL"
-    )
-    href = blocks.URLBlock(
-        required=False, help_text="Pick a page or specify a URL", label="URL"
-    )
-    size = BootstrapButtonSizeChoiceBlock(required=False, default="md")
-    style = BootstrapButtonStyleChoiceBlock(required=False, default="btn-outline-dark")
-
-    class Meta:
-        template = "app/blocks/cta.html"
-
-
-class HeroTextBlock(blocks.StructBlock):
-    heading = blocks.CharBlock(max_length=250, form_classname="full title")
-    background_color = BackgroundColourChoiceBlock(required=False)
-    button = ButtonBlock(required=False)
-
-    class Meta:
-        template = "app/blocks/hero_block.html"
-        icon = "fa fa-alphabet"
-
-
-class ListItemBlock(blocks.StructBlock):
-    title = blocks.CharBlock(max_length=250, form_classname="full title")
-    image = ImageChooserBlock(required=False)
-    image_css = blocks.CharBlock(max_length=500, required=False)
-    caption = blocks.RichTextBlock(max_length=500)
-    background_color = BackgroundColourChoiceBlock(required=False)
-    button = ButtonBlock(required=False)
-
-    class Meta:
-        template = "app/blocks/list_item_block.html"
-        icon = "fa fa-alphabet"
-
-
-class ColumnWidthChoiceBlock(blocks.ChoiceBlock):
-    choices = [
-        ("small", "small"),
-        ("medium", "medium"),
-        ("large", "large"),
-    ]
-
-    class Meta:
-        icon = "fa-arrows-alt"
-        default = "small"
-
-
-class ListBlock(blocks.StructBlock):
-    column_width = ColumnWidthChoiceBlock()
-    items = blocks.ListBlock(ListItemBlock)
-
-    class Meta:
-        template = "app/blocks/list_block.html"
-        icon = "fa fa-alphabet"
-
-
-class FeaturedBookBlock(blocks.StructBlock):
-    book = blocks.PageChooserBlock(
-        page_type="app.bookpage",
-        target_model="app.bookpage",
-        can_choose_root=False,
-    )
-    background_color = BackgroundColourChoiceBlock(required=False)
-    promotion_label = blocks.CharBlock(
-        required=False, help_text="Label that highlights this product"
-    )
-    description = blocks.RichTextBlock(
-        required=False,
-        help_text="This will replace the book's default description. You can use this to provide a more contextualised description of the book",
-    )
-
-    class Meta:
-        template = "app/blocks/featured_book_block.html"
-        icon = "fa fa-book"
-
-
-class FeaturedProductBlock(blocks.StructBlock):
-    product = blocks.PageChooserBlock(
-        page_type="app.merchandisepage",
-        target_model="app.merchandisepage",
-        can_choose_root=False,
-    )
-    background_color = BackgroundColourChoiceBlock(required=False)
-    promotion_label = blocks.CharBlock(
-        required=False, help_text="Label that highlights this product"
-    )
-    description = blocks.RichTextBlock(
-        required=False,
-        help_text="This will replace the product's default description. You can use this to provide a more contextualised description of the product",
-    )
-
-    class Meta:
-        template = "app/blocks/featured_product_block.html"
-        icon = "fa fa-shopping-cart"
-
-
-class BookGridBlock(blocks.StructBlock):
-    column_width = ColumnWidthChoiceBlock()
-
-    class Meta:
-        template = "app/blocks/book_grid_block.html"
-        icon = "fa fa-book"
-
-
-class ProductGridBlock(blocks.StructBlock):
-    column_width = ColumnWidthChoiceBlock()
-
-    class Meta:
-        template = "app/blocks/product_grid_block.html"
-        icon = "fa fa-shopping-cart"
-
-
-class SelectedBooksBlock(BookGridBlock):
-    books = blocks.ListBlock(
-        blocks.PageChooserBlock(
-            page_type="app.bookpage",
-            target_model="app.bookpage",
-            can_choose_root=False,
-        )
-    )
-
-    def get_context(self, value, parent_context=None):
-        context = super().get_context(value, parent_context)
-        context["books"] = value["books"]
-        return context
-
-
-class SelectedProductsBlock(ProductGridBlock):
-    products = blocks.ListBlock(
-        blocks.PageChooserBlock(
-            page_type="app.merchandisepage",
-            target_model="app.merchandisepage",
-            can_choose_root=False,
-        )
-    )
-
-    def get_context(self, value, parent_context=None):
-        context = super().get_context(value, parent_context)
-        context["products"] = list(value["products"])
-        return context
-
-
-class RecentlyPublishedBooks(BookGridBlock):
-    max_books = blocks.IntegerBlock(
-        default=4, help_text="How many books should show up?"
-    )
-    type = BookTypeChoice(default="all-books")
-
-    def get_context(self, value, parent_context=None):
-        context = super().get_context(value, parent_context)
-        filters = {}
-        if value["type"] != None and value["type"] != "all-books":
-            filters["type"] = value["type"]
-        context["books"] = (
-            BookPage.objects.order_by("-published_date")
-            .live()
-            .public()
-            .filter(published_date__isnull=False, **filters)
-            .all()[: value["max_books"]]
-        )
-        return context
-
-
-class FullProductList(ProductGridBlock):
-    max_products = blocks.IntegerBlock(
-        default=10, help_text="How many products should show up?"
-    )
-
-    def get_context(self, value, parent_context=None):
-        context = super().get_context(value, parent_context)
-        context["products"] = list(
-            MerchandisePage.objects.live().public().all()[: value["max_products"]]
-        )
-        return context
-
-
-class YourBooks(BookGridBlock):
-    class Meta:
-        template = "app/blocks/your_books_grid_block.html"
-        icon = "fa fa-book"
-
-
-class SingleBookBlock(blocks.StructBlock):
-    book = blocks.PageChooserBlock(
-        page_type="app.bookpage",
-        target_model="app.bookpage",
-        can_choose_root=False,
-    )
-
-    class Meta:
-        template = "app/blocks/single_book_block.html"
-        icon = "fa fa-book"
-
-
-class NewsletterSignupBlock(blocks.StructBlock):
-    heading = blocks.CharBlock(required=False, max_length=150)
-
-    class Meta:
-        template = "app/blocks/newsletter_signup_block.html"
-        icon = "fa fa-email"
-
-
-class ArticleText(blocks.StructBlock):
-    text = blocks.RichTextBlock(form_classname="full", features=block_features)
-    alignment = AlignmentChoiceBlock(
-        help_text="Doesn't apply when used inside a column."
-    )
-
-    class Meta:
-        template = "app/blocks/richtext.html"
-
-
-class ColumnBlock(blocks.StructBlock):
-    stream_blocks = [
-        ("hero_text", HeroTextBlock()),
-        ("title_image_caption", ListItemBlock()),
-        ("image", ImageChooserBlock()),
-        ("single_book", SingleBookBlock()),
-        ("membership_plan", PlanBlock()),
-        ("richtext", blocks.RichTextBlock(features=block_features)),
-        ("button", ButtonBlock()),
-        ("newsletter_signup", NewsletterSignupBlock()),
-    ]
-    background_color = BackgroundColourChoiceBlock(required=False)
-    content = blocks.StreamBlock(stream_blocks, required=False)
-
-
-class SingleColumnBlock(ColumnBlock):
-    column_width = ColumnWidthChoiceBlock()
-    alignment = AlignmentChoiceBlock(
-        help_text="Doesn't apply when used inside a column."
-    )
-
-    class Meta:
-        template = "app/blocks/single_column_block.html"
-        icon = "fa fa-th-large"
-
-
-class MultiColumnBlock(blocks.StructBlock):
-    background_color = BackgroundColourChoiceBlock(required=False)
-    columns = blocks.ListBlock(ColumnBlock, min_num=1, max_num=5)
-
-    class Meta:
-        template = "app/blocks/columns_block.html"
-        icon = "fa fa-th-large"
-
-
-class EventsListBlock(blocks.StructBlock):
-    number_of_events = blocks.IntegerBlock(required=True, default=3)
-
-    def get_context(self, value, parent_context=None):
-        context = super().get_context(value, parent_context)
-        context.update(MapPage.get_map_context())
-        return context
-
-    class Meta:
-        template = "app/blocks/event_list_block.html"
-        icon = "fa fa-calendar"
-
-
-class EventsListAndMap(blocks.StructBlock):
-    title = blocks.CharBlock(required=False)
-    intro = blocks.RichTextBlock(required=False)
-    number_of_events = blocks.IntegerBlock(required=True, default=3)
-
-    def get_context(self, value, parent_context=None):
-        context = super().get_context(value, parent_context)
-        context.update(MapPage.get_map_context())
-        return context
-
-    class Meta:
-        template = "app/blocks/event_list_and_map_block.html"
-        icon = "fa fa-map"
-
-
-def create_streamfield(additional_blocks=None, **kwargs):
-    blcks = [
-        ("membership_options", MembershipOptionsBlock()),
-        ("image", ImageChooserBlock()),
-        ("featured_book", FeaturedBookBlock()),
-        ("book_selection", SelectedBooksBlock()),
-        ("recently_published_books", RecentlyPublishedBooks()),
-        ("featured_product", FeaturedProductBlock()),
-        ("product_selection", SelectedProductsBlock()),
-        ("full_product_list", FullProductList()),
-        ("hero_text", HeroTextBlock()),
-        ("heading", blocks.CharBlock(form_classname="full title")),
-        ("richtext", ArticleText()),
-        ("list_of_heading_image_text", ListBlock()),
-        ("single_column", SingleColumnBlock()),
-        ("columns", MultiColumnBlock()),
-        ("events_list_and_map", EventsListAndMap()),
-        ("events_list_block", EventsListBlock()),
-    ]
-
-    if isinstance(additional_blocks, list):
-        blcks += additional_blocks
-
-    return StreamField(blcks, null=True, blank=True, use_json_field=True, **kwargs)
-
-
 @method_decorator(cache_page, name="serve")
 class MerchandiseIndexPage(WagtailCacheMixin, IndexPageSeoMixin, Page):
     show_in_menus_default = True
@@ -1286,10 +869,11 @@ class MembershipPlanPage(WagtailCacheMixin, ArticleSeoMixin, Page):
     parent_page_types = ["app.HomePage"]
 
     ### v2 signup flow fields
-    display_in_quiz_flow = models.BooleanField(default=False)
+    display_in_quiz_flow = models.BooleanField(
+        default=False, verbose_name="Display in v2 signup flow"
+    )
     benefits = RichTextField(
-        features=["ul"],
-        help_text="List of pithy beneficial features of this plan",
+        help_text="List of pithy beneficial features of this plan. Bullet points are automatically formatted as ticks.",
         null=True,
         blank=True,
     )
@@ -1317,13 +901,21 @@ class MembershipPlanPage(WagtailCacheMixin, ArticleSeoMixin, Page):
         blank=True,
     )
 
+    @classmethod
+    def create_default_layout(cls):
+        return (
+            {"id": uuid.uuid4(), "type": "plan_title", "value": {}},
+            {"id": uuid.uuid4(), "type": "plan_pricing", "value": {}},
+        )
+
     layout = create_streamfield(
         [("plan_title", PlanTitleBlock()), ("plan_pricing", PlanPricingBlock())],
-        default=(("plan_title", {}), ("plan_pricing", {})),
+        default=create_default_layout,
     )
 
     panels = content_panels = Page.content_panels + [
         FieldPanel("deliveries_per_year", heading="[v1+v2] Deliveries per year"),
+        InlinePanel("prices", min_num=1, label="Prices"),
         MultiFieldPanel(
             [
                 FieldPanel("display_in_quiz_flow"),
@@ -1336,7 +928,6 @@ class MembershipPlanPage(WagtailCacheMixin, ArticleSeoMixin, Page):
         ),
         MultiFieldPanel(
             [
-                InlinePanel("prices", min_num=1, label="Prices"),
                 InlinePanel("upsells", label="Upsell prices"),
                 FieldPanel("pick_product_title", classname="full title"),
                 FieldPanel("pick_product_text"),
@@ -1380,7 +971,9 @@ class MembershipPlanPage(WagtailCacheMixin, ArticleSeoMixin, Page):
         return self.prices.filter(interval="year").order_by("interval_count").first()
 
     @property
-    def annual_percent_off_per_month(self) -> str:
+    def annual_percent_off_per_month(self):
+        if self.basic_price is None or self.annual_price is None:
+            return None
         return (
             self.annual_price.equivalent_monthly_price - self.basic_price.price
         ) / self.basic_price.price
