+++ app/templates/app/blocks/syllabus_title_block.html
@@ -1,0 +1,10 @@
+{% load wagtailcore_tags %}
+<header class='bg-white py-3 py-lg-5 mb-lg-4'>
+    <div class='row'>
+        <div class='col-12 text-center'>
+            <div class='opacity-50'>Syllabus</div>
+            <h1>{{ page.title }}</h1>
+            {% if page.description %}<div class='h5'>{{ page.description|richtext }}</div>{% endif %}
+        </div>
+    </div>
+</header>
