+++ app/templates/app/signup/select_deliveries.html
@@ -9,6 +9,7 @@
         {% comment %}  {% endcomment %}
         <div class="row justify-content-center tw-flex">
             <form data-turbo="false" method="POST" class="tw-mx-auto tw-max-w-lg">
+                <h2 class='h4 fw-bold text-center tw-mb-4'>How often do you want a new book delivered?</h2>
                 {% csrf_token %}
                 <ul class="tw-grid tw-w-full tw-gap-2 tw-list-none tw-p-0">
                     {% for option in delivery_options %}
@@ -21,11 +22,16 @@
                                    class="tw-hidden tw-peer"
                                    required/>
                             <label for="option-{{ option.id }}"
-                                   class="tw-inline-flex tw-items-center tw-justify-between tw-w-full tw-p-5 tw-text-gray-500 tw-bg-white tw-border tw-border-gray-200 tw-rounded-lg tw-cursor-pointer peer-checked:tw-border-blue-600 peer-checked:tw-text-blue-600 hover:tw-text-gray-600 hover:tw-bg-gray-100">
+                                   class="tw-inline-flex tw-items-center tw-justify-between tw-w-full tw-p-4 tw-text-gray-500 tw-bg-white tw-border tw-border-gray-200 tw-rounded-lg tw-cursor-pointer peer-checked:tw-border-blue-600 peer-checked:tw-text-blue-600 hover:tw-text-gray-600 hover:tw-bg-gray-100">
                                 <div class="tw-block">
-                                    <div class="tw-w-full tw-text-lg tw-font-semibold">{{ option.title }}</div>
-                                    <div class="tw-w-full">{{ option.monthly_price.price_string_uk_pp }} including shipping</div>
-                                    <div class="tw-w-full">{{ option.description|richtext }}</div>
+                                    <div class="tw-w-full tw-text-lg tw-font-semibold tw-text-black">{{ option.title }}</div>
+                                    <div class="tw-w-full">
+                                        {{ option.monthly_price.monthly_price_string }} + {{ option.monthly_price.shipping_price_string_uk }} shipping
+                                    </div>
+                                    <div class="tw-w-full -tw-mb-3">{{ option.description|richtext }}</div>
+                                    <div class="tw-w-full [&>ul]:tw-pl-4 tw-list-image-[url('data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iMTQiIGhlaWdodD0iMTIiIHZpZXdCb3g9IjAgMCAxNCAxMiIgeG1sbnM9Imh0dHA6Ly93d3cudzMub3JnLzIwMDAvc3ZnIiBmaWxsPSIjMzhiZGY4Ij48cGF0aCBmaWxsLXJ1bGU9ImV2ZW5vZGQiIGQ9Ik0xMy42ODUuMTUzYS43NTIuNzUyIDAgMCAxIC4xNDMgMS4wNTJsLTggMTAuNWEuNzUuNzUgMCAwIDEtMS4xMjcuMDc1bC00LjUtNC41YS43NS43NSAwIDAgMSAxLjA2LTEuMDZsMy44OTQgMy44OTMgNy40OC05LjgxN2EuNzUuNzUgMCAwIDEgMS4wNS0uMTQzWiIgLz48L3N2Zz4=')]">
+                                        {{ option.benefits|richtext }}
+                                    </div>
                                 </div>
                                 <svg class="tw-w-5 tw-h-5 tw-ms-3 rtl:tw-rotate-180"
                                      aria-hidden="true"
