+++ app/templates/app/signup/select_donation.html
@@ -8,7 +8,20 @@
             {% signup_header steps=steps %}
             <div class="row justify-content-center">
                 <section class='col-12 col-md-6 col-lg-4 mb-5'>
-                    <div class='max-width-card mx-auto'>{% include "app/includes/donation_form.html" with confirm_text="Continue" %}</div>
+                    <div class='max-width-card mx-auto'>
+                        <h2 class='h4 fw-bold text-center'>Can you add a {{ price.interval }}ly donation?</h2>
+                        <p class='tw-mb-4 text-center'>
+                            <small>
+                                <em>
+                                    <strong>
+                                        <a href='/about' target='_blank' rel='noopener noreferrer'>LBC is a not-for-profit political education project.</a>
+                                    </strong>
+                                    We rely on donations from our supporters to keep our membership affordable to all.
+                                </em>
+                            </small>
+                        </p>
+                        {% include "app/includes/donation_form.html" with confirm_text="Continue" continue_without=True %}
+                    </div>
                 </section>
             </div>
         {% endwith %}
