+++ app/templates/app/syllabus_page.html
@@ -1,0 +1,1 @@
+{% extends "layout_page.html" %}
