+++ app/templates/components/signup_header.html
@@ -1,12 +1,11 @@
 {% with request.GET.gift_mode as gift_mode %}
-    <header class="pricing-header my-4 my-lg-5 text-center">
-        <h1 class="h2 fw-bold text-center">
+    <header class="pricing-header my-3 my-lg-4 text-center">
+        <h1 class="h3 tw-font-normal text-center">
             {% comment %} {% if gift_mode %}
-            You're buying a membership gift card
-        {% else %}
-            You're joining the Left Book Club
-            {% endif %} {% endcomment %}
-            Design your membership
+                Gift a Left Book Club membership
+        {% else %} {% endcomment %}
+            Join the Left Book Club
+            {% comment %} {% endif %} {% endcomment %}
         </h1>
         <nav style="--bs-breadcrumb-divider: url(&#34;
                     data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='8' height='8'%3E%3Cpath d='M2.5 0L1 1.5 3.5 4 1 6.5 2.5 8l4-4-4-4z' fill='currentColor'/%3E%3C/svg%3E&#34;
