+++ frontend/controllers/donation-controller.ts
@@ -3,15 +3,13 @@ import { Controller } from "@hotwired/stimulus";
 class DonationController extends Controller {
   static targets = ["input"];
 
-  async update(e: InputEvent) {
+  setInputs(value: number) {
     // If the new value is one of the radios, select them
     // Else unselect all radios
-
-    const value = (e.target as HTMLInputElement)?.value;
     if (value) {
       this.inputTargets.forEach((input) => {
         if (input.type === "radio") {
-          input.checked = parseFloat(input.value) === parseFloat(value);
+          input.checked = parseFloat(input.value) === value;
         } else {
           input.value = value;
         }
@@ -21,10 +19,20 @@ class DonationController extends Controller {
         if (input.type === "radio") {
           input.checked = false;
         } else {
-          input.value = "";
+          input.value = "0";
         }
       });
     }
+  }
+
+  update(e: InputEvent) {
+    let value = parseFloat((e.target as HTMLInputElement)?.value || "0") || 0;
+    this.setInputs(value);
+  }
+
+  async submitWithoutDonation(e: InputEvent) {
+    this.setInputs(0);
+    (this.element as HTMLFormElement).submit();
   }
 }
